+++ include/executor/task_handlers/instagram.hpp
@@ -56,7 +56,7 @@ class IGTaskHandler : public TaskHandler {
     std::vector<FileInfo> task_files = parseFileInfo(file_info);
 
     std::string media_filename = get_executable_cwd();
-    for (int i = 0; i < task_files.size(); i++) {
+    for (uint8_t i = 0; i < task_files.size(); i++) {
       task_files.at(i).first =
           media_filename + "/data/" + uuid + "/" + task_files.at(i).first;
     }
